+++ python/GafferImageUI/LUTUI.py
@@ -90,8 +90,8 @@ Gaffer.Metadata.registerNode(
 			""",
 
 			"plugValueWidget:type", "GafferUI.PresetsPlugValueWidget",
-			"preset:Forward", GafferImage.LUT.Direction.Forward,
-			"preset:Inverse", GafferImage.LUT.Direction.Inverse,
+			"preset:Forward", GafferImage.OpenColorIOTransform.Direction.Forward,
+			"preset:Inverse", GafferImage.OpenColorIOTransform.Direction.Inverse,
 
 		],